+++ components/Nav.tsx
@@ -1,37 +1,27 @@
 "use client";
+
 import { useGlobalData } from "@/hooks/useGlobalData";
 import { faBars } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { AnimatePresence } from "framer-motion";
-import { useRouter } from "next/navigation";
 import React, { useState } from "react";
+import { PillButton } from "./PillButton";
 import { SmallScreenNav } from "./SmallScreenNav";
 import { WaitListModal } from "./WaitListModal";
 
 export const Nav = () => {
   const [showSmallScreenMenu, setShowSmallScreenMenu] = useState(false);
 
-  const { closeFn, isShowingWaitlistModal, openWaitlistModal } =
-    useGlobalData();
+  const { handleStart, isShowingWaitlistModal } = useGlobalData();
 
   const hideSmallScreenMenu = () => setShowSmallScreenMenu(false);
-
-  const router = useRouter();
-  const { user } = useGlobalData();
-  const handleStart = () => {
-    if (user) {
-      router.push("/dashboard");
-    } else {
-      router.push("/auth");
-    }
-  };
 
   return (
     <nav className="flex px-[5%] py-[2%] bg-gtahidiCream sticky top-0 right-0 left-0 z-50">
       <ul className="sm:ml-auto flex gap-x-4 text-sm items-center w-full sm:w-max justify-between sm:justify-normal">
         <FontAwesomeIcon
           icon={faBars}
-          className="h-5 w-5 text-gtahidiPurple block sm:hidden"
+          className="h-5 w-5 text-gtahidiPurple hidden"
           onClick={() => setShowSmallScreenMenu(true)}
         />
         <AnimatePresence>
@@ -53,12 +43,7 @@ export const Nav = () => {
             Pricing
           </a>
         </div>
-        <button
-          className="bg-gtahidiPurple p-3 rounded text-white font-medium"
-          onClick={handleStart}
-        >
-          Start Creating
-        </button>
+        <PillButton onClick={handleStart}>Start Creating</PillButton>
       </ul>
       {isShowingWaitlistModal && <WaitListModal />}
     </nav>
