+++ components/StartCreating.tsx
@@ -3,7 +3,7 @@ import { useGlobalData } from "@/hooks/useGlobalData";
 import React from "react";
 
 export const StartCreating = () => {
-  const { openWaitlistModal } = useGlobalData();
+  const { handleStart } = useGlobalData();
   return (
     <section className="p-[5%] flex justify-around items-center flex-col sm:flex-row">
       <p className="text-4xl font-semibold">
@@ -11,9 +11,9 @@ export const StartCreating = () => {
       </p>
       <button
         className="bg-gtahidiPurple py-4 px-6 hover:scale-125 transition-all rounded-full text-white font-medium mt-5 sm:mt-0"
-        onClick={openWaitlistModal}
+        onClick={handleStart}
       >
-        Join the waitlist
+        Start creating
       </button>
     </section>
   );
