+++ components/Welcome.tsx
@@ -5,7 +5,7 @@ import Image from "next/image";
 import React from "react";
 
 export const Welcome = () => {
-  const { openWaitlistModal } = useGlobalData();
+  const { handleStart } = useGlobalData();
   return (
     <header className="p-[5%] flex bg-gtahidiPurpleTwo justify-between flex-col sm:flex-row">
       <h1 className="text-3xl w-full sm:w-1/4">
@@ -21,9 +21,9 @@ export const Welcome = () => {
         </p>
         <button
           className="py-5 bg-gtahidiPurple rounded-lg text-white w-1/2"
-          onClick={openWaitlistModal}
+          onClick={handleStart}
         >
-          Join the waitlist
+          Start Creating
         </button>
       </div>
       <Image
