+++ hooks/useGlobalData.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useRouter } from "next/navigation";
 import React from "react";
 import { createContext, useContext, useState } from "react";
 import { User } from "../types";
@@ -11,16 +12,14 @@ export const DEFAULT_PRICE = {
 };
 
 type GlobalData = {
-  closeFn: () => void;
   isShowingWaitlistModal: boolean;
-  openWaitlistModal: () => void;
+  handleStart: () => void;
   user: User | null;
 };
 
 export const globalDataContext = createContext<GlobalData>({
-  closeFn: () => {},
   isShowingWaitlistModal: false,
-  openWaitlistModal: () => {},
+  handleStart: () => {},
   user: null,
 });
 
@@ -32,20 +31,21 @@ export const GlobalDataProvider = ({
   const [isShowingWaitlistModal, setIsShowWaitlistModal] = useState(false);
   const [user, setUser] = useState<User | null>(null);
 
-  const closeFn = () => {
-    setIsShowWaitlistModal(false);
-  };
+  const router = useRouter();
 
-  const openWaitlistModal = () => {
-    setIsShowWaitlistModal(true);
+  const handleStart = () => {
+    if (user) {
+      router.push("/dashboard");
+    } else {
+      router.push("/auth");
+    }
   };
 
   return (
     <globalDataContext.Provider
       value={{
         isShowingWaitlistModal,
-        closeFn,
-        openWaitlistModal,
+        handleStart,
         user,
       }}
     >
